+++ code_test/src/backjoon/july/No24264.java
@@ -1,4 +1,4 @@
-package backjoon.july5;
+package backjoon.july;
 
 import java.util.Scanner;
