+++ code_test/src/backjoon/july/No24266.java
@@ -1,0 +1,16 @@
+package backjoon.july;
+
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.InputStreamReader;
+
+public class No24266 {
+    public static void main(String[] args) throws IOException {
+        BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
+
+        long a = Long.parseLong(br.readLine());
+
+        System.out.println(a*a*a);
+        System.out.println(3);
+    }
+}
