+++ code_test/src/backjoon/july/No24313.java
@@ -1,0 +1,5 @@
+package backjoon.july;
+
+public class No24313 {
+    
+}
